+++ src/App.js
@@ -11,13 +11,27 @@ import Posts from "./components/Posts";
 import Profile from "./components/Profile";
 import CreatePost from "./components/CreatePost";
 import Post from "./components/Post";
+import CategoryService from "./services/category.service";
 
 const App = () => {
   const [currentUser, setCurrentUser] = useState(undefined);
+  const [categories, setCategories] = useState([]);
+
+  const retrieveCategories = () => {
+    CategoryService.getAllCategories()
+        .then(response => {
+          setCategories(...response.data);
+          console.log("test");
+          console.log(response.data);
+        })
+        .catch(e => {
+          console.log(e);
+        });
+  };
 
   useEffect(() => {
     const user = AuthService.getCurrentUser();
-
+    retrieveCategories();
     if (user) {
       setCurrentUser(user);
     }
@@ -29,7 +43,7 @@ const App = () => {
 
   return (
       <div>
-        <nav className="navbar navbar-expand navbar-dark bg-dark">
+        <nav className="navbar navbar-expand navbar-dark bg-dark nav">
           <Link to={"/"} className="navbar-brand">
             Uni-Platform
           </Link>
@@ -80,8 +94,8 @@ const App = () => {
           <Routes>
             <Route path="/" element={<Posts/>} />
             <Route path="/posts" element={<Posts/>} />
-            <Route path="/posts/create" element={<CreatePost/>} />
-            <Route path="/posts/:id" element = {<Post/>} />
+            <Route path="/posts/create" element={<CreatePost categories={categories}/>} />
+            <Route path="/posts/:id" element = {<Post categories={categories}/>} />
             <Route path="/login" element={<Login/>} />
             <Route path="/register" element={<Register/>} />
             <Route path="/profile" element={<Profile/>} />
